refactor: move api services to providers
