+++ studio/src/app/providers/api/photo/api.photo.factory.service.ts
@@ -1,5 +1,5 @@
 import {EnvironmentAppConfig} from '../../../types/core/environment-config';
-import {EnvironmentConfigService} from '../../environment/environment-config.service';
+import {EnvironmentConfigService} from '../../../services/environment/environment-config.service';
 
 import {ApiPhotoService} from './api.photo.service';
 import {ApiPhotoMockService} from './api.photo.mock.service';
