+++ studio/src/app/providers/api/photo/api.photo.mock.service.ts
