+++ studio/src/app/providers/api/photo/api.photo.service.ts
